Move Show/Hide more results notification into a separate file
