+++ src/app/pages/SearchResultsPage/ScopedSearchResultsView.tsx
@@ -1,9 +1,5 @@
 import { FC, useState } from 'react'
-import { Trans, useTranslation } from 'react-i18next'
-import { styled } from '@mui/material/styles'
-import Box from '@mui/material/Box'
-import ZoomIn from '@mui/icons-material/ZoomIn'
-import ZoomOut from '@mui/icons-material/ZoomOut'
+import { useTranslation } from 'react-i18next'
 import {
   getFilterForLayer,
   getFilterForNetwork,
@@ -23,30 +19,7 @@ import { SearchResults } from './hooks'
 import { SearchResultsList } from './SearchResultsList'
 import { NoResults } from './NoResults'
 import { AllTokenPrices } from '../../../coin-gecko/api'
-
-const NotificationBox = styled(Box)(({ theme }) => ({
-  marginTop: 30,
-  marginBottom: 30,
-  fontSize: 14,
-  marginLeft: '10%',
-  marginRight: '10%',
-
-  boxSizing: 'border-box',
-  display: 'flex',
-  flexDirection: 'row',
-  justifyContent: 'center',
-  alignItems: 'center',
-  padding: '5px 10px',
-  gap: 10,
-
-  height: 50,
-
-  background: theme.palette.layout.notificationBackground,
-  border: `2px solid ${theme.palette.layout.darkBorder}`,
-  color: theme.palette.layout.notificationText,
-
-  borderRadius: 50,
-}))
+import { HideMoreResults, ShowMoreResults } from './notifications'
 
 export const ScopedSearchResultsView: FC<{
   wantedScope: SearchScope
@@ -79,12 +52,7 @@ export const ScopedSearchResultsView: FC<{
       )}
       {othersOpen ? (
         <>
-          <NotificationBox theme={notificationTheme} onClick={() => setOthersOpen(false)}>
-            <ZoomOut />
-            <span>
-              <Trans i18nKey={'search.otherResults.clickToHide'} />
-            </span>
-          </NotificationBox>
+          <HideMoreResults theme={notificationTheme} onHide={() => setOthersOpen(false)} />
           {RouteUtils.getEnabledNetworks()
             .filter(net => net !== wantedScope.network)
             .map(net => (
@@ -99,24 +67,12 @@ export const ScopedSearchResultsView: FC<{
         </>
       ) : (
         !!otherResults.length && (
-          <NotificationBox theme={notificationTheme} onClick={() => setOthersOpen(true)}>
-            <ZoomIn />
-            <span>
-              <Trans
-                t={t}
-                i18nKey={
-                  hasWantedResults
-                    ? 'search.otherResults.clickToShowMore'
-                    : 'search.otherResults.clickToShowThem'
-                }
-                values={{
-                  countLabel: t(hasWantedResults ? 'search.results.moreCount' : 'search.results.count', {
-                    count: otherResults.length,
-                  }),
-                }}
-              />
-            </span>
-          </NotificationBox>
+          <ShowMoreResults
+            theme={notificationTheme}
+            onShow={() => setOthersOpen(true)}
+            hasWantedResults={hasWantedResults}
+            otherResultsCount={otherResults.length}
+          />
         )
       )}
     </>
